requirements/common.txt
Django==4.0
python-decouple==3.5
django-tenants==3.4.1
psycopg2-binary==2.9.3
boto3==1.20.32
django-storages==1.12.3
python-keycloak==0.26.1
djangorestframework==3.13.1
django-filter==21.1
djangorestframework-simplejwt==5.0.0
celery==5.2.3
redis==4.1.1
python-magic==0.4.25
django-redis==5.2.0
paypalrestsdk==1.13.1
Pillow==9.0.1
django-cors-headers==3.11.0
africastalking==1.1.8
